+++ pyh-nv/pages/pyh-nv/transparent/transparent.vue
@@ -12,7 +12,7 @@
 		data() {
 			return {
 				nvConfig:{
-					title:" ",
+					title:"pyh-nv 渐变导航栏",
 					bgColor:"#ffffff",
 					// componentBgColor:"#2b9939",
 					color:"#000000",